+++ app/build.gradle
@@ -10,15 +10,14 @@ plugins {
 }
 
 android {
-    compileSdkVersion 32
-    buildToolsVersion "30.0.3"
+    compileSdk 34
 
     defaultConfig {
         applicationId "com.desuzed.everyweather"
         minSdkVersion 21
-        targetSdkVersion 32
-        versionCode 6
-        versionName "2.1.0"
+        targetSdkVersion 34
+        versionCode 7
+        versionName "2.1.1"
         javaCompileOptions {
             annotationProcessorOptions {
                 arguments = ["room.schemaLocation":
@@ -82,8 +81,8 @@ dependencies {
     implementation 'com.google.android.material:material:1.4.0'
 
     testImplementation 'junit:junit:4.+'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
     //compose
     implementation "com.google.maps.android:maps-compose:1.0.0"
