+++ app/src/main/java/com/desuzed/everyweather/domain/model/settings/SettingUiItem.kt
@@ -1,0 +1,6 @@
+package com.desuzed.everyweather.domain.model.settings
+
+class SettingUiItem<S : Enum<*>>(
+    val setting: S,
+    val textId: Int,
+)
