+++ app/src/main/java/com/desuzed/everyweather/presentation/features/settings/SettingsViewModel.kt
@@ -6,7 +6,17 @@ import com.desuzed.everyweather.data.repository.local.SettingsDataStore
 import com.desuzed.everyweather.data.repository.providers.app_update.AppUpdateProvider
 import com.desuzed.everyweather.domain.model.app_update.AppUpdateState
 import com.desuzed.everyweather.domain.model.app_update.InAppUpdateStatus
-import com.desuzed.everyweather.domain.model.settings.*
+import com.desuzed.everyweather.domain.model.settings.DarkMode
+import com.desuzed.everyweather.domain.model.settings.DarkTheme
+import com.desuzed.everyweather.domain.model.settings.DistanceDimen
+import com.desuzed.everyweather.domain.model.settings.Lang
+import com.desuzed.everyweather.domain.model.settings.Language
+import com.desuzed.everyweather.domain.model.settings.Pressure
+import com.desuzed.everyweather.domain.model.settings.PressureDimen
+import com.desuzed.everyweather.domain.model.settings.SettingsType
+import com.desuzed.everyweather.domain.model.settings.TempDimen
+import com.desuzed.everyweather.domain.model.settings.Temperature
+import com.desuzed.everyweather.domain.model.settings.WindSpeed
 import com.desuzed.everyweather.presentation.base.BaseViewModel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
@@ -24,6 +34,7 @@ class SettingsViewModel(
         collect(settingsDataStore.tempDimen, ::collectTemperature)
         collect(settingsDataStore.pressureDimen, ::collectPressure)
         collect(appUpdateProvider.appUpdateState, ::onAppUpdateState)
+        initSettingItemsLists()
     }
 
     override fun onUserInteraction(interaction: SettingsUserInteraction) {
@@ -120,14 +131,28 @@ class SettingsViewModel(
         copy(pressure = pressure)
     }
 
+    private fun initSettingItemsLists() {
+        setState {
+            copy(
+                langDialogItems = settingsDataStore.getLanguageItemsList(),
+                darkModeDialogItems = settingsDataStore.getDarkModeItemsList(),
+                temperatureDialogItems = settingsDataStore.getTemperatureItemsList(),
+                distanceDialogItems = settingsDataStore.getDistanceItemsList(),
+                pressureDialogItems = settingsDataStore.getPressureItemsList(),
+            )
+        }
+    }
+
     private fun onAppUpdateState(appUpdateState: AppUpdateState?) {
         when (appUpdateState) {
             AppUpdateState.UpdateAvailable -> {
                 setState { copy(updateStatus = InAppUpdateStatus.READY_TO_LAUNCH_UPDATE) }
             }
+
             AppUpdateState.ReadyToInstall -> {
                 setState { copy(updateStatus = InAppUpdateStatus.READY_TO_INSTALL) }
             }
+
             else -> {
                 setState { copy(updateStatus = null) }
             }
