+++ app/src/main/java/com/desuzed/everyweather/presentation/features/settings/ui/SettingsMenuGroupContent.kt
@@ -1,0 +1,78 @@
+package com.desuzed.everyweather.presentation.features.settings.ui
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Icon
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import com.desuzed.everyweather.R
+import com.desuzed.everyweather.domain.model.settings.BaseSettingItem
+import com.desuzed.everyweather.presentation.features.settings.SettingsUserInteraction
+import com.desuzed.everyweather.ui.elements.BoldText
+import com.desuzed.everyweather.ui.elements.MediumText
+import com.desuzed.everyweather.ui.elements.RoundedCardItem
+import com.desuzed.everyweather.ui.theming.EveryweatherTheme
+
+@Composable
+fun SettingsMenuGroupContent(
+    modifier: Modifier = Modifier,
+    items: List<BaseSettingItem>,
+    onUserInteraction: (SettingsUserInteraction) -> Unit
+) {
+    val interactionSource = remember { MutableInteractionSource() }
+    RoundedCardItem(
+        modifier = Modifier.padding(dimensionResource(id = R.dimen.dimen_10))
+    ) {
+        Column(modifier = Modifier.padding(dimensionResource(id = R.dimen.dimen_10))) {
+            items.forEach { item ->
+                val onItemCLick: () -> Unit = {
+                    onUserInteraction(
+                        SettingsUserInteraction.ShowSettingDialog(item.type)
+                    )
+                }
+                Row(
+                    modifier = Modifier
+                        .padding(
+                            horizontal = dimensionResource(id = R.dimen.dimen_10),
+                            vertical = dimensionResource(id = R.dimen.dimen_8)
+                        )
+                        .fillMaxWidth()
+                        .clickable(
+                            interactionSource = interactionSource,
+                            indication = null,
+                            onClick = onItemCLick,
+                        ),
+                    horizontalArrangement = Arrangement.SpaceBetween,
+                    verticalAlignment = Alignment.CenterVertically,
+
+                    ) {
+                    Column {
+                        BoldText(
+                            text = stringResource(id = item.categoryStringId),
+                            onClick = onItemCLick
+                        )
+                        MediumText(text = stringResource(item.valueStringId))
+                    }
+                    Icon(
+                        modifier = Modifier.rotate(90f),
+                        painter = painterResource(id = R.drawable.ic_arrow_24),
+                        contentDescription = "",
+                        tint = EveryweatherTheme.colors.textColorPrimary
+                    )
+                }
+            }
+
+        }
+    }
+}
