+++ app/src/main/java/com/desuzed/everyweather/presentation/features/weather_next_days/NextDaysBottomSheet.kt
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.platform.ComposeView
+import com.desuzed.everyweather.presentation.features.weather_next_days.ui.NextDaysBottomSheetContent
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
