+++ app/src/main/java/com/desuzed/everyweather/ui/Previews.kt
@@ -1,0 +1,10 @@
+package com.desuzed.everyweather.ui
+
+import android.content.res.Configuration
+import androidx.compose.ui.tooling.preview.Preview
+
+@Preview(
+    showBackground = true,
+    uiMode = Configuration.UI_MODE_NIGHT_NO,
+)
+annotation class AppPreview
